requirements.txt
conda==4.6.14
conda-build==3.15.1
Cython==0.28.5
cytoolz==0.9.0.1
dicom-numpy==0.1.6
h5py==2.8.0
html5lib==1.0.1
image==1.5.27
ImageContour==0.2.5.0
imageio==2.4.1
imagesize==1.1.0
imutils==0.5.1
ipython==6.5.0
ipython-genutils==0.2.0
Keras==2.2.4
Keras-Applications==1.0.8
keras-contrib==2.0.8
Keras-Preprocessing==1.1.0
matplotlib==3.0.3
nibabel==2.3.1
nilearn==0.5.2
np-utils==0.5.9.0
numpy==1.17.0
numpydoc==0.8.0
pandas==0.23.4
pandocfilters==1.4.2
pickleshare==0.7.4
Pillow==5.4.1
pillowfight==0.3
plotly==4.1.1
pydicom==1.2.2
scikit-image==0.15.0
scikit-learn==0.19.2
scipy==1.2.1
seaborn==0.9.0
SimpleITK==1.2.0
statsmodels==0.9.0
tables==3.4.4
tensorboard==1.14.0
tensorflow==1.13.1
tensorflow-estimator==1.14.0rc0
tensorflow-gpu==1.14.0